requirements.txt
fastapi
uvicorn[standard]
python-jose[cryptography]
pandas
python-multipart